Back to old db structure
finally worked out how to build the query to get the correct users, using an indexedQuery.
now it first creates a map of all users, sets that to the event then writes the entire event.
when useing dataSnapshot.getValue(Event.class) the invited users map is now included.

also fixed pullToRefresh to dismiss uring onDataChanged of the adapters.